backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
alembic
pytest
httpx
debugpy
prometheus-fastapi-instrumentator
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-sqlalchemy